feat(booth) Basic translation of booth Added Spanish translation files for booth UI to locale directory.
Changed settings.py: Changed default locale to es-ES, indicated list of
available languages (English and Spanish for now) and location of translation
files (locale directory).

For now, the booth will be displayed in English if the user's browser requests
the page in English, in all other cases it will be displayed in Spanish, which
is the default language.